MERGEDPR-01:features/CRUDOperation
* CHANGED: CRUDOperation with SQL

* CHANGED: Project files

* CHANGED: CRUDOperation UniqueISBN

---------

+++ BookManagement.Web/PuthagaUlagam.Common/ValidationMethods/InputValidation.cs
@@ -21,7 +21,7 @@ namespace PuthagaUlagam.Common
 
         public static bool IsValidPrice(string Price)
         {
-            if (!int.TryParse(Price, out _))
+            if (!decimal.TryParse(Price, out _))
             {
                 return false;
             }
